chore: Move generateTransactionsTableColumns
