+++ apps/explorer/src/lib/ui/utils/generateTransactionsTableColumns.tsx
@@ -7,7 +7,7 @@ import type { IotaTransactionBlockKind, IotaTransactionBlockResponse } from '@io
 
 import { TableCellBase, TableCellText } from '@iota/apps-ui-kit';
 import type { ColumnDef } from '@tanstack/react-table';
-import { AddressLink, ObjectLink } from '../ui';
+import { AddressLink, ObjectLink } from '../../../components/ui';
 
 /**
  * Generate table columns renderers for the transactions data.
