+++ apps/explorer/src/lib/ui/utils/index.ts
@@ -4,4 +4,5 @@
 export * from './generateCheckpointsTableColumns';
 export * from './generateEpochsTableColumns';
 export * from './generateValidatorsTableColumns';
+export * from './generateTransactionsTableColumns';
 export * from './objectField';
